feat: organize helper functions into utils package
